(hybris) load system properties before "on ready" trigger. JB#39724
This is to avoid libhybris clients loading the wrong linker during boot,
because ro.build.version.sdk (which libhybris checks for) is undefined
causing SIGSEGV on processes that are started very early in the boot process.

We may load system props that early, since /system is already mounted by systemd.

Signed-off-by: Franz-Josef Haider <f_haider@gmx.at>